Add base code for decompressing archives.
The commit should give an idea about how the code and the protocol
between JavaScript and NaCl are designed.
Most of the logic for the JavaScript code is finished (with the
exception of reading files from the archive).
For now the NaCL code contains only fake metadata that will be sent
back to Javascript. The next commit should retrieve real metadata from
the archives.

Review: https://team-review.git.corp.google.com/#/c/6930/

BUG=chromium:388077
Change-Id: I4fa715c74c74a8ad148af8d50f4aa6be731ba028