Modify metadata in order to use a hierarchy structure instead of a flat structure.
Use a hierarchy structure instead of a flat structure for metadata. The
change is done directly from NaCl. The reason is that in case of lots of
entries in a folder with a long name, the actual data might be less to send.
e.g.
"/",
"/longDirectoryName",
"/longDirectoryName/file1",
"/longDirectoryName/file2",
......................
"/longDirectoryName/file1000"

The new structure will be
entries: {
  "longDirectoryName": {
    entries: {
      "file1",
      "file2",
      .....
      "file1000"
    }
  }
}
As we can see 1000 "longDirectoryName" were changed to a simple
dictionary. The old structure was useful only if there were lots and
lots of folders on the first and second level of the dictionary.

Also improves search speeds for onGetMetadataRequested and
onReadDirectoryRequested.

Review: https://team-review.git.corp.google.com/7161

BUG=chromium:388077
Change-Id: I50f3d4be42c8a494c595ae04eb3ae4f1ce2603e9