Read real metadata for zip and rar archives.
Obtain real metadata for zip and rar archives using libarchive.
Add instructions to README on how to use libarchive. Modify Makefile in
order to build using libarchive.

Add the NaCl base code structure that will be used by future CLs.
The NaCl module uses libarchive to read metadata. Modify JavaScript files in
order to read the archive file and send it to NaCl for processing.
Repair tests. The workflow is explained below.

When a user opens a new RAR archive in the Files App, the call gets
dispatched to the Chrome extension. The Chrome extension will send a
Read Metadata request to NaCl, which will request from libarchive one by
one the metadata for every entry in the archive (this is how libarchive
works). Libarchive contains the algorithms for obtaining the metadata
from an archive’s data, but for that I had to send the archive’s data to
libarchive.

The read is done using 2 threads: one that blocks for processing an
entry’s metadata until archive’s data is available (libarchive
requirement), and one that reads the actual archive’s data in Javascript
and unblocks the other thread when this data is ready to be processed.

Once all entries are obtained, the result is returned to the Files App,
which can display the archive’s contents to the user.

BUG=chromium:388077
Change-Id: Ib00b2924703bed6207fc8c79454c5aee7acde9a6