Fix error message in decompressor_test.js and global definitions in JS tests.
Fix error message in decompressor_test.js in order to suggest that tests
pass even if an error is printed. This is the expected behavior because
the error execution is tested.
Move global definitions for all JS test files to 'describe' function, except
for integration_test.js. This is required as karma might use a global
definition from another file. integration_test.js will be fixed in
another CL as most of its logic will be changed.

BUG=chromium:407003
Change-Id: I69fd1babc1dfd219fab646034026023700b71d61