Solve restore state bug after restarts and suspend.
In case of multiple calls made to the fileSystemProvider API after
restart / suspend the extension would fail to load the metadata for the
archives. The reason is that the calls would lead to races as they don't
wait for the previous calls to end.

This CL solves this issue by using DOM events and queueing all the calls
related to a file system. The calls are queued only per file system as
using a single queue would lead to deadlocks in case of an archive in an
archive (inner archive will wait for data from outer archive, while
outer archive will wait for inner archive restore call to end).

BUG=chromium:388077
Change-Id: I6453d570fe063227543bd6d78e6d0ee2300def68