Resolve a memory corruption bug due to ReadAhead.
Until ReadAhead, volume reading was thread safe, but ReadAhead can request
data that will not be used anymore in the future. Due to this reason
that data can be received either a CloseFile event that removes the
corresponding VolumeArchive / VolumeReader from the
worker_reads_in_progress_ map which can lead to races.

Also in case we receive a ReadAhead with the same read_offset as the
VolumeReader offset then we can end up with memory corruption.
This was possible because the use of the buffer with current data until
switching the buffers for reading ahead was not guarded. Current CL resolves
this issue by ignoring read ahead data that is received after receiving
first read ahead buffer. There is still no need to guard that zone as we
cannot end up with buffer overwrite anymore.

BUG=chromium:388077
Change-Id: Idb4eb8f7babe6d950394a620a86ef5305d9077ba