Force unmount for failed restores and save state only for the affected volumes.
In case a volume fails to restore after suspend / restart then force
its unmount. The volume is now in an inconsistent state. This failure
can lead to inconsistent local storage for other volumes if state save is done
for all volumes at the same time during the forced unmount.
So save and remove state should be done only for the affected volumes.

BUG=chromium:388077
Change-Id: I3fabc69af00f98ede6183ef269cd5efeeb14c49c