Unmount on passphrase failure.
If a passphrase dialog is cancelled, then unmount the archive, as otherwise
it becomes unusable (accessing files causes errors).

Note, that ideally cancelling a passphrase dialog should only cancel it for the
file being accessed, and accessing other files should invoke the dialog again.
This is however impossible for now due to a bug in libarchive.

TEST=Tested manually.
BUG=None

Change-Id: Ia72d5044ac18f66180e4a064e0f64031809c33a9