Fix unmounting archives when the source file is gone.
The previous logic worked if the extension was restoring. However if it's
already loaded, then reading would fail without unmounting.

This CL fixes the issue by adding an unmounting call when reading a chunk
fails due to a (possibly) missing file.

Also, startup logic is removed, as it's not needed to be handled anymore.
Some remaining code caused a JS error on startup, as the registered method
was undefined.

TEST=Tested manually.
BUG=None

Change-Id: I0339abe6e338afdb3a1b08b88ff8f765a86467ef