WGU: support compressed (but not archived) files
People sometimes get compressed files like foo.gz or foo.bz2.
They are not archived at all, so the existing archive logic does
not handle them properly.

Rework the layers so that we fallback to the "raw" format in
libarchive when the compression+archive filters could not find
a match.  This allows us to run it through compression filters
only.

We have to fake out metadata fields a bit more ourselves too.

Change-Id: Idfa2e94c949d04fc8b9637a2af09817253ca6849