Unload NaCL module when it's not necessary.
Unload the NaCL module when there is no mounted volume and no ongoing
mounting process. To check whether a mounting process is going on,
the counter is added that represents the number of mounting process.
This counter is incremented and decremented in EnsureVolumeLoaded()
and onLaunched(), which are the start points of mounting process.

BUG=517741
TEST=manually tested

Change-Id: I9274d445cd36bfd4c0fe551641cecf4fbbf09c9d