tests: update docs & avoid global installs
Since we install all the npm modules in the top level, we can
use the tools installed directly in that rather than having to
install tools into the host OS.

We also simplify the test docs to use a single npm install.
With current versions, we also have to manually install the
chai and sinon packages, so add them to the list.

Change-Id: Ib197ba15f26c46d62513f5c36f54621c5d636b6f