tests: link to gtest source files directly
Upstream gtest does not provide a libgtest.a anymore to link against.
Instead, they want you to compile it from source all the time when you
use them with your own projects.

Adapt the test makefile to use the gtest source in the NaCl SDK, and
have it build gtest-all.cc directly instead of using libgtest.a.

Change-Id: I0aa087c35d32a6446402a5d33f487c6877471c84