do not rely on file size if it isn't available
In some cases, libarchive might not have been able to set the file size
for the current entry.  If it couldn't, we get a default of 0 bytes, and
this reading code would return early.

Instead, if the size isn't available, just rely on the libarchive layers
returning EOF on the current member once we read to the end of it.

Change-Id: I89d223c1d68dd7d6bd05335dd76a3490eb0ffdef