Always update buffer and send signal when ReadChunk is done.
There is a chance that VolumeReaderJavaScriptStream::Read() waits a response of
ReadChunk forever.
The problematic scenario is illustrated in
https://bugs.chromium.org/p/chromium/issues/detail?id=457448#c28

VolumeReaderJavaScriptStream can request ReadChunk without a client's request to
read ahead.
It expects a response through VolumeReaderJavaScriptStream::SetBufferAndSignal,
but a response can be ignored in Volume::Volume::ReadChunkDone() if the file for
previous Read request is already closed. This can cause Read() to be hanged.

Considering the case of reading ahead (without explicit Read requests from
client), I think we should always set buffer and signal
VolumeReaderJavaScriptStream when ReadChunk is done.

BUG=chromium:457448
TEST=Open a zip file which contains 1000 photos, and browse photos by
scrolling randomly.
Change-Id: I4850b89ee6d5f8e808e0c94f76aabb4a3eda35b7