normalize archive paths
Some archive formats allow names to include "./" prefixes.  These confuse
the FSP layer and make it list a "." directory.  Attempting to descend it
though fail as the FSP layer requests "./" which libarchive returns which
is the same directory it was already in.

Lets change the unpacker layer to just skip any leading ./ components it
comes across.  The FSP layer can then display it fine while the libarchive
layer knows how to look up the path even with the ./ omitted.

An example using a tar file:
$ touch foo
$ tar cf foo.tar ./foo
Opening it in the FSP layer shows a hidden "." directory, and going into
it leads to an infinite ./././././ chain that never reaches "foo".

We can also confuse it with runs of // or /./ in dirpaths, so normalize
those too.

Bug: https://github.com/vapier/chrome-ext-wicked-good-unarchiver/issues/7
Change-Id: Ieefd9a769dd840b1029e5c1d773db2eec1cc74ca