WGU: update to latest libarchive code
Rebase the zip work that we've been doing in CrOS to the latest
libarchive tree.  We pull it out into the nacl patch so we can
point directly to the upstream libarchive repo.

Everything has been merged upstream at this point other than the
seek support that we use with ZIP files.

Change-Id: I9b261021c76699687af34f4176f40c94fd277e9f